+++ packages/sdks/output/nextjs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Builder.io Next.js SDK Changelog (@builder.io/sdk-react-nextjs)
 
+## 0.17.7
+
+### Patch Changes
+
+- abe5cba: Feat: exports `setClientUserAttributes` helper that can be used to set and update Builder's user attributes cookie. This cookie is used by Builder's Personalization Containers to decide which variant to render.
+
+  Usage example:
+
+  ```ts
+  import { setClientUserAttributes } from "@builder.io/sdk-react-nextjs";
+
+  setClientUserAttributes({
+    device: "tablet",
+  });
+  ```
+
 ## 0.17.6
 
 ### Patch Changes