+++ packages/sdks/output/solid/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Builder.io Solid SDK Changelog (@builder.io/sdk-solid)
 
+## 3.0.6
+
+### Patch Changes
+
+- abe5cba: Feat: exports `setClientUserAttributes` helper that can be used to set and update Builder's user attributes cookie. This cookie is used by Builder's Personalization Containers to decide which variant to render.
+
+  Usage example:
+
+  ```ts
+  import { setClientUserAttributes } from "@builder.io/sdk-solid";
+
+  setClientUserAttributes({
+    device: "tablet",
+  });
+  ```
+
 ## 3.0.5
 
 ### Patch Changes